RAC: Refactor front structure
