+++ src/Akeneo/Platform/Job/front/process-tracker/src/feature/components/JobExecutionDetail/Progress.test.tsx
