+++ src/Akeneo/Platform/Job/front/process-tracker/src/feature/components/JobExecutionDetail/index.ts
@@ -1,0 +1,3 @@
+export * from './Progress';
+export * from './ShowProfile';
+export * from './summary';
