+++ src/Akeneo/Platform/Job/front/process-tracker/src/feature/components/JobExecutionDetail/summary/InnerTable.test.tsx
