+++ src/Akeneo/Platform/Job/front/process-tracker/src/feature/components/JobExecutionDetail/summary/index.ts
@@ -1,0 +1,3 @@
+export * from './InnerTable';
+export * from './SummaryTable';
+export * from './WarningHelper';
