+++ src/Akeneo/Platform/Job/front/process-tracker/src/feature/components/JobExecutionList/JobExecutionTable.tsx
@@ -2,8 +2,7 @@ import React, {useCallback, MouseEvent} from 'react';
 import {Table} from 'akeneo-design-system';
 import {useDateFormatter, useRouter, useSecurity, useTranslate} from '@akeneo-pim-community/shared';
 import {JobExecutionRow, JobExecutionFilterSort, jobCanBeStopped, canShowJobExecutionDetail} from '../../models';
-import {JobExecutionStatus} from '../JobExecutionStatus';
-import {StopJobAction} from '../StopJobAction';
+import {JobExecutionStatus, StopJobAction} from '../common';
 
 type JobExecutionTableProps = {
   sticky?: number;
