+++ src/Akeneo/Platform/Job/front/process-tracker/src/feature/components/LastExecutionTable.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import {renderWithProviders} from '@akeneo-pim-community/shared';
 import {act, screen, within} from '@testing-library/react';
 import {LastExecutionTable} from './LastExecutionTable';
-import {JobExecutionRow} from 'feature/models/JobExecutionTable';
+import {JobExecutionRow} from '../models';
 import userEvent from '@testing-library/user-event';
 
 const rows: JobExecutionRow[] = [
