+++ src/Akeneo/Platform/Job/front/process-tracker/src/feature/components/LastExecutionTable.tsx
@@ -2,8 +2,7 @@ import React, {MouseEvent} from 'react';
 import {Table} from 'akeneo-design-system';
 import {useDateFormatter, useRouter, useSecurity, useTranslate} from '@akeneo-pim-community/shared';
 import {JobExecutionRow, jobCanBeStopped, canShowJobExecutionDetail} from '../models';
-import {JobExecutionStatus} from './JobExecutionStatus';
-import {StopJobAction} from './StopJobAction';
+import {JobExecutionStatus, StopJobAction} from './common';
 
 const COLUMN_HEADERS = ['started_at', 'username', 'status', 'warning_count'];
