+++ src/Akeneo/Platform/Job/front/process-tracker/src/feature/components/common/JobExecutionStatus.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {Badge, Level} from 'akeneo-design-system';
 import {useTranslate} from '@akeneo-pim-community/shared';
-import {JobStatus} from '../models';
+import {JobStatus} from '../../models';
 
 const badgeLevel = (status: JobStatus, hasError: boolean, hasWarning: boolean): Level => {
   if (status === 'FAILED' || hasError) {
