+++ src/Akeneo/Platform/Job/front/process-tracker/src/feature/components/common/index.ts
@@ -1,0 +1,2 @@
+export * from './JobExecutionStatus';
+export * from './StopJobAction';
