+++ src/Akeneo/Platform/Job/front/process-tracker/src/feature/components/index.ts
@@ -1,3 +1,4 @@
+export * from './common';
 export * from './JobExecutionList';
-export * from './JobExecutionStatus';
+export * from './JobExecutionDetail';
 export * from './LastExecutionTable';
