+++ src/Akeneo/Platform/Job/front/process-tracker/src/feature/hooks/useJobExecution.test.ts
@@ -65,6 +65,7 @@ const successResponse = {
 };
 
 test('It returns the fetched job execution', async () => {
+  jest.useFakeTimers('modern');
   global.fetch = jest.fn().mockImplementation(async () => ({
     ok: true,
     json: async () => successResponse,
@@ -72,6 +73,7 @@ test('It returns the fetched job execution', async () => {
 
   const {result, waitForNextUpdate} = renderHookWithProviders(() => useJobExecution('1'));
   await act(async () => {
+    jest.advanceTimersByTime(1000);
     await waitForNextUpdate();
   });
 
@@ -82,6 +84,7 @@ test('It returns the fetched job execution', async () => {
 });
 
 test('It returns error when fetch return an error', async () => {
+  jest.useFakeTimers('modern');
   global.fetch = jest.fn().mockImplementation(async () => ({
     ok: false,
     statusText: 'Not found',
@@ -90,6 +93,7 @@ test('It returns error when fetch return an error', async () => {
 
   const {result, waitForNextUpdate} = renderHookWithProviders(() => useJobExecution('1'));
   await act(async () => {
+    jest.advanceTimersByTime(1000);
     await waitForNextUpdate();
   });
 
@@ -103,6 +107,7 @@ test('It returns error when fetch return an error', async () => {
 });
 
 test('It returns callback to reload job execution information', async () => {
+  jest.useFakeTimers('modern');
   global.fetch = jest.fn().mockImplementation(async () => ({
     ok: true,
     json: async () => successResponse,
@@ -110,6 +115,7 @@ test('It returns callback to reload job execution information', async () => {
 
   const {result, waitForNextUpdate} = renderHookWithProviders(() => useJobExecution('1'));
   await act(async () => {
+    jest.advanceTimersByTime(1000);
     await waitForNextUpdate();
   });
 
@@ -161,6 +167,7 @@ test('It returns callback to reload job execution information', async () => {
 });
 
 test('It does not fetch a job execution while the previous fetch is not finished', async () => {
+  jest.useFakeTimers('modern');
   global.fetch = jest.fn().mockImplementation(
     async () =>
       new Promise(resolve =>
@@ -177,6 +184,7 @@ test('It does not fetch a job execution while the previous fetch is not finished
 
   const {result, waitForNextUpdate} = renderHookWithProviders(() => useJobExecution('1'));
   await act(async () => {
+    jest.advanceTimersByTime(1000);
     await waitForNextUpdate();
   });
