+++ src/Akeneo/Platform/Job/front/process-tracker/src/feature/models/index.ts
@@ -1,3 +1,4 @@
 export * from './JobExecutionFilter';
 export * from './JobExecutionTable';
+export * from './JobExecutionDetail';
 export * from './JobStatus';
